requirements.txt
flask==3.0.2
google-generativeai==0.8.4
flask-cors==4.0.0
